updates to update command switches

+++ install/runcheckout.sh
@@ -297,18 +297,23 @@ for module in ${javamodules[@]}; do
 				fi
 			fi
   else
-    cd $module
-	printf "\e[32m$module: \e[0m"
-    sudo git reset --hard
-    sudo git fetch --all
-    sudo git checkout tags/$branch 2>/dev/null
-    if [ ! $? = 0 ]; then sudo git checkout $branch 2>/dev/null; fi
-	if [ ! $? = 0 ]; then echo "no branch $branch exists, switching to $defaultbranch"; sudo git checkout $defaultbranch 2>/dev/null; fi
 
-	## fast forward to latest head
-	if [ ! "$nopull" = "1" ]; then
-		echo "Pulling latest changes: "
-		sudo git pull
+	cd $module
+
+	if [ -d ".git" ]; then 
+		printf "\e[32m$module: \e[0m"
+	    sudo git reset --hard
+	    sudo git fetch --all
+	    sudo git checkout tags/$branch 2>/dev/null
+	    if [ ! $? = 0 ]; then sudo git checkout $branch 2>/dev/null; fi
+		if [ ! $? = 0 ]; then echo "no branch $branch exists, switching to $defaultbranch"; sudo git checkout $defaultbranch 2>/dev/null; fi
+
+		## fast forward to latest head
+		if [ ! "$nopull" = "1" ]; then
+			echo "Pulling latest changes: "
+			sudo git pull
+		fi
+
 	fi
 
     cd ..
